requirements.txt
numpy
pandas
scikit-learn
scikit-image
flask
streamlit
keras
tensorflow
mtcnn
torch
pydot
graphviz
seaborn
matplotlib 
xgboost
PyYAML
boto3
pytest